+++ controllers/usuario.js
@@ -64,18 +64,17 @@ const getUsuarioPorId = async (req = request, res = response) => {
  */
 const updateUsuarioPorId = async (req = request, res = response) => {
     try {
-        const { id } = req.params
-        const { nombre, email, contrasena, direccion, esVendedor } = req.body
-        const data = {
-            nombre,
-            email,
-            contrasena,
-            direccion,
-            esVendedor
-        }
-        const usuario = 
-            await Usuario.findByIdAndUpdate(id, data, {new : true});
-        res.status(201).json(usuario)
+
+        const { id } = req.query
+        const data = req.body
+
+        const usuarioDB = await Usuario.findByIdAndUpdate(id, data, {new : true});
+
+        if(!usuarioDB) return res.json({msg: 'No hay datos'})
+        
+        return res.json({usuarioDB})
+
+
     }catch(e) {
         return res.status(500).json({msj: e})
     }
@@ -86,9 +85,18 @@ const updateUsuarioPorId = async (req = request, res = response) => {
  */
 const deleteUsuarioByID = async (req = request, res = response) => {
     try{
-        const id = req.params.id;
-        const usuario = await Usuario.findByIdAndDelete(id);
-        res.status(204).json(usuario);
+            const {id} = req.query
+
+            const usuarioDB = await Usuario.findById(id)
+
+            if(usuarioDB){
+                const usuarioDBfound = await Usuario.findByIdAndDelete(id)
+                return res.json({msg: 'El usuario fue eliminado con exito'})
+            }
+            if(!usuarioDB){
+                return res.json({msg: 'No existe ese id'})
+            } 
+        
          }catch(e) {
             return res.status(500).json({msj: e})
         }
